feat: 모임리스트 #22 feat: 상세페이지 #14
